# This is a combination of 2 commits. # The first commit's message is:
fix rerun of socket port

When using puppet it always returns returns as sync, the value of port
is an int in the file on disk but in the catalouge it's a string. This
commit uses insync?(is) to always convert socket[:port] to an integer

# This is the 2nd commit message:

Added additional API user and password parameters.